Moved individual validations in separate files. Removed useless comments.
