+++ Alignment/OfflineValidation/python/TkAlAllInOneTool/genericValidation.py
@@ -1,0 +1,96 @@
+import os
+
+class GenericValidation:
+    defaultReferenceName = "DEFAULT"
+    def __init__(self, valName, alignment, config):
+        import random
+        self.name = valName
+        self.alignmentToValidate = alignment
+        self.general = config.getGeneral()
+        self.randomWorkdirPart = "%0i"%random.randint(1,10e9)
+        self.configFiles = []
+        self.filesToCompare = {}
+        self.jobmode = self.general["jobmode"]
+        # check, if it has advantages to include the config as validation member
+        self.config = config
+
+    def getRepMap(self, alignment = None):
+        if alignment == None:
+            alignment = self.alignmentToValidate
+        result = alignment.getRepMap()
+        result.update( self.general )
+        result.update({
+                "workdir": os.path.join( self.general["workdir"],
+                                         self.randomWorkdirPart ),
+                "datadir": self.general["datadir"],
+                "logdir": self.general["logdir"],
+                "dbLoad": alignment.getLoadTemplate(),
+                "APE": alignment.getAPETemplate(),
+                "CommandLineTemplate": ("#run configfile and post-proccess it\n"
+                                        "cmsRun %(cfgFile)s\n"
+                                        "%(postProcess)s "),
+                "CMSSW_BASE": os.environ['CMSSW_BASE'],
+                "SCRAM_ARCH": os.environ['SCRAM_ARCH'],
+                "alignmentName": alignment.name,
+                "condLoad": alignment.getConditions()
+                })
+        return result
+
+    def getCompareStrings( self, requestId = None ):
+        result = {}
+        repMap = self.alignmentToValidate.getRepMap()
+        for validationId in self.filesToCompare:
+            repMap["file"] = self.filesToCompare[ validationId ]
+            if repMap["file"].startswith( "/castor/" ):
+                repMap["file"] = "rfio:%(file)s"%repMap
+            result[validationId]= "%(file)s=%(name)s|%(color)s|%(style)s"%repMap
+        if requestId == None:
+            return result
+        else:
+            if not "." in requestId:
+                requestId += ".%s"%GenericValidation.defaultReferenceName
+            if not requestId.split(".")[-1] in result:
+                msg = ("could not find %s in reference Objects!"
+                       %requestId.split(".")[-1])
+                raise AllInOneError(msg)
+            return result[ requestId.split(".")[-1] ]
+
+    def createFiles( self, fileContents, path ):
+        result = []
+        for fileName in fileContents:
+            filePath = os.path.join( path, fileName)
+            theFile = open( filePath, "w" )
+            theFile.write( fileContents[ fileName ] )
+            theFile.close()
+            result.append( filePath )
+        return result
+
+    def createConfiguration(self, fileContents, path, schedule= None):
+        self.configFiles = GenericValidation.createFiles(self, fileContents,
+                                                         path) 
+        if not schedule == None:
+            schedule = [os.path.join( path, cfgName) for cfgName in schedule]
+            for cfgName in schedule:
+                if not cfgName in self.configFiles:
+                    msg = ("scheduled %s missing in generated configfiles: %s"
+                           %(cfgName, self.configFiles))
+                    raise AllInOneError(msg)
+            for cfgName in self.configFiles:
+                if not cfgName in schedule:
+                    msg = ("generated configuration %s not scheduled: %s"
+                           %(cfgName, schedule))
+                    raise AllInOneError(msg)
+            self.configFiles = schedule
+        return self.configFiles
+
+    def createScript(self, fileContents, path, downloadFiles=[] ):        
+        self.scriptFiles = GenericValidation.createFiles(self, fileContents,
+                                                         path)
+        for script in self.scriptFiles:
+            os.chmod(script,0755)
+        return self.scriptFiles
+
+    def createCrabCfg(self, fileContents, path ):        
+        self.crabConfigFiles = GenericValidation.createFiles(self, fileContents,
+                                                             path)
+        return self.crabConfigFiles
