+++ Alignment/OfflineValidation/python/TkAlAllInOneTool/globalDictionaries.py
@@ -1,0 +1,5 @@
+# Store used datasets, to avoid making the same DAS query multiple times
+usedDatasets = {}
+
+# Needed for more than one geometry comparison for one alignment
+alignRandDict = {}
