+++ Alignment/OfflineValidation/python/TkAlAllInOneTool/zMuMuValidation.py
@@ -1,0 +1,147 @@
+import os
+import configTemplates
+import globalDictionaries
+import dataset as datasetModule
+from genericValidation import GenericValidation
+from helperFunctions import replaceByMap
+
+
+class ZMuMuValidation(GenericValidation):
+    def __init__(self, valName, alignment,config):
+        GenericValidation.__init__(self, valName, alignment, config)
+        defaults = {
+            "zmumureference": ("/afs/cern.ch/cms/CAF/CMSALCA/ALCA_TRACKERALIGN2"
+                               "/TMP_EM/ZMuMu/data/MC/BiasCheck_DYToMuMu_Summer"
+                               "11_TkAlZMuMu_IDEAL.root"),
+            "jobmode":self.jobmode,
+            "runRange":"",
+            "firstRun":"",
+            "lastRun":"",
+            "begin":"",
+            "end":"",
+            "JSON":""
+            }
+        mandatories = [ "dataset", "maxevents",
+#                         "etamax1", "etamin1", "etamax2", "etamin2" ]
+                        "etamaxneg", "etaminneg", "etamaxpos", "etaminpos" ]
+        if not config.has_section( "zmumu:"+self.name ):
+            zmumu = config.getResultingSection( "general",
+                                                  defaultDict = defaults,
+                                                  demandPars = mandatories )
+        else:
+            zmumu = config.getResultingSection( "zmumu:"+self.name, 
+                                                  defaultDict = defaults,
+                                                  demandPars = mandatories )
+        self.general.update( zmumu )
+        self.jobmode = self.general["jobmode"]
+        if self.general["dataset"] not in globalDictionaries.usedDatasets:
+            globalDictionaries.usedDatasets[self.general["dataset"]] = datasetModule.Dataset(
+                self.general["dataset"] )
+        self.dataset = globalDictionaries.usedDatasets[self.general["dataset"]]
+    
+    def createConfiguration(self, path, configBaseName = "TkAlZMuMuValidation" ):
+        cfgName = "%s.%s.%s_cfg.py"%( configBaseName, self.name,
+                                      self.alignmentToValidate.name )
+        repMap = self.getRepMap()
+        cfgs = {cfgName:replaceByMap( configTemplates.ZMuMuValidationTemplate, repMap)}
+        GenericValidation.createConfiguration(self, cfgs, path)
+        
+    def createScript(self, path, scriptBaseName = "TkAlZMuMuValidation"):
+        scriptName = "%s.%s.%s.sh"%(scriptBaseName, self.name,
+                                    self.alignmentToValidate.name )
+        repMap = self.getRepMap()
+        repMap["CommandLine"]=""
+        for cfg in self.configFiles:
+            repMap["CommandLine"]+= repMap["CommandLineTemplate"]%{"cfgFile":cfg,
+                                                  "postProcess":""
+                                                  }
+        scripts = {scriptName: replaceByMap( configTemplates.zMuMuScriptTemplate, repMap ) }
+        return GenericValidation.createScript(self, scripts, path)
+
+    def createCrabCfg( self, path,
+                       crabCfgBaseName = "TkAlZMuMuValidation"  ):
+        crabCfgName = "crab.%s.%s.%s.cfg"%( crabCfgBaseName, self.name,
+                                            self.alignmentToValidate.name )
+        repMap = self.getRepMap()
+        repMap["script"] = "dummy_script.sh"
+        repMap["crabOutputDir"] = os.path.basename( path )
+        repMap["crabWorkingDir"] = crabCfgName.split( '.cfg' )[0]
+        self.crabWorkingDir = repMap["crabWorkingDir"]
+        repMap["numberOfJobs"] = self.general["parallelJobs"]
+        repMap["cfgFile"] = self.configFiles[0]
+        repMap["queue"] = self.jobmode.split( ',' )[1].split( '-q' )[1]
+        if self.dataset.dataType() == "mc":
+            repMap["McOrData"] = "events = .oO[nEvents]Oo."
+        elif self.dataset.dataType() == "data":
+            repMap["McOrData"] = "lumis = -1"
+            if self.jobmode.split( ',' )[0] == "crab":
+                print ("For jobmode 'crab' the parameter 'maxevents' will be "
+                       "ignored and all events will be processed.")
+        crabCfg = {crabCfgName: replaceByMap( configTemplates.crabCfgTemplate,
+                                              repMap ) }
+        return GenericValidation.createCrabCfg( self, crabCfg, path )
+
+    def getRepMap(self, alignment = None):
+        repMap = GenericValidation.getRepMap(self, alignment) 
+        repMap.update({
+            "nEvents": self.general["maxevents"],
+#             "outputFile": "zmumuHisto.root"
+            "outputFile": ("0_zmumuHisto.root"
+                           ",genSimRecoPlots.root"
+                           ",FitParameters.txt")
+                })
+        if not self.jobmode.split( ',' )[0] == "crab":
+            try:
+                repMap["datasetDefinition"] = self.dataset.datasetSnippet(
+                    jsonPath = self.general["JSON"],
+                    nEvents = self.general["maxevents"],
+                    firstRun = self.general["firstRun"],
+                    lastRun = self.general["lastRun"],
+                    begin = self.general["begin"],
+                    end = self.general["end"] )
+            except AllInOneError, e:
+                msg = "In section [zmumu:%s]: "%( self.name )
+                msg += str( e )
+                raise AllInOneError( msg )
+        else:
+            if self.dataset.predefined():
+                msg = ("For jobmode 'crab' you cannot use predefined datasets "
+                       "(in your case: '%s')."%( self.dataset.name() ))
+                raise AllInOneError( msg )
+            if self.general["begin"] or self.general["end"]:
+                ( self.general["firstRun"],
+                  self.general["lastRun"] ) = self.dataset.convertTimeToRun(
+                    firstRun = self.general["firstRun"],
+                    lastRun = self.general["lastRun"],
+                    begin = self.general["begin"],
+                    end = self.general["end"] )
+                self.general["firstRun"] = str( self.general["firstRun"] )
+                self.general["lastRun"] = str( self.general["lastRun"] )
+            if ( not self.general["firstRun"] ) and \
+                   ( self.general["end"] or self.general["lastRun"] ):
+                self.general["firstRun"] = str( self.dataset.runList()[0]["run_number"] )
+            if ( not self.general["lastRun"] ) and \
+                   ( self.general["begin"] or self.general["firstRun"] ):
+                self.general["lastRun"] = str( self.dataset.runList()[-1]["run_number"] )
+            if self.general["firstRun"] and self.general["lastRun"]:
+                if int( self.general["firstRun"] ) > int( self.general["lastRun"] ):
+                    msg = ( "The lower time/runrange limit ('begin'/'firstRun') "
+                            "chosen is greater than the upper time/runrange limit "
+                            "('end'/'lastRun').")
+                    raise AllInOneError( msg )
+                self.general["runRange"] = (self.general["firstRun"]
+                                            + '-' + self.general["lastRun"])
+            try:
+                repMap["datasetDefinition"] = self.dataset.datasetSnippet(
+                    jsonPath = self.general["JSON"],
+                    nEvents = self.general["maxevents"],
+                    firstRun = self.general["firstRun"],
+                    lastRun = self.general["lastRun"],
+                    begin = self.general["begin"],
+                    end = self.general["end"],
+                    crab = True )
+            except AllInOneError, e:
+                msg = "In section [zmumu:%s]: "%( self.name )
+                msg += str( e )
+                raise AllInOneError( msg )
+        return repMap
